--- src/entities/create-form/ui/MultipleChoiceOption/index.tsx
@@ -1,32 +1,0 @@
-import React from 'react';
-import { Circle } from '@/shared/assets/svg';
-import { OptionProps } from '@/shared/types/create-form/type';
-import OptionItem from '../OptionItem';
-import OtherOption from '../OtherOptionProps';
-
-const MultipleChoiceOption = ({
-  fields,
-  remove,
-  register,
-  index,
-  isCheckBox,
-}: OptionProps) => {
-  return (
-    <div className="space-y-2">
-      {fields.map((option, optionIndex) => (
-        <OptionItem
-          key={option.id}
-          icon={<Circle />}
-          optionId={option.id}
-          optionIndex={optionIndex}
-          register={register}
-          remove={remove}
-          inputName={`questions.${index}.options.${optionIndex}.value`}
-        />
-      ))}
-      {isCheckBox ? <OtherOption text="기타" /> : null}
-    </div>
-  );
-};
-
-export default MultipleChoiceOption;
