+++ src/views/form/create/api/createForm.ts
@@ -1,0 +1,13 @@
+import axios from 'axios';
+import { CreateFormRequest } from '@/shared/types/form/create/type';
+
+export const createForm = async ({
+  data,
+  id,
+}: {
+  data: CreateFormRequest;
+  id: string;
+}) => {
+  const response = await axios.post(`/api/form/${id}`, data);
+  return response;
+};
